bus_tracking/backend/requirements.txt
# Django Backend Dependencies
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.0.0

# Database (optional - for production)
# psycopg2-binary>=2.9.9  # PostgreSQL
# mysqlclient>=2.2.0      # MySQL

# Development utilities
python-decouple>=3.8
django-extensions>=3.2.3

# PDF Generation
reportlab>=4.0.0
xhtml2pdf>=0.2.11

# Testing
pytest>=7.4.0
pytest-django>=4.5.0

# Production server
gunicorn>=21.2.0
whitenoise>=6.6.0
